move bias related logic to tinygemm

+++ test/srt/cpu/test_moe.py
@@ -8,6 +8,7 @@ kernel = torch.ops.sgl_kernel
 
 torch.manual_seed(1234)
 
+from sglang.test.test_utils import CustomTestCase
 from utils import (
     BLOCK_K,
     BLOCK_N,
@@ -23,8 +24,6 @@ from utils import (
     torch_naive_fused_moe_gptoss,
     torch_w8a8_per_column_fused_moe,
 )
-
-from sglang.test.test_utils import CustomTestCase
 
 
 def fused_moe(a, w1, w2, score, topk, renormalize, prepack):
@@ -362,8 +361,8 @@ class TestFusedExperts(CustomTestCase):
             None,  # block_size
             None,  # a1_scale
             None,  # a2_scale
-            w1_b,
-            w2_b,
+            w1_b.float(),
+            w2_b.float(),
             alpha,
             limit,
             True,  # is_vnni
